+++ src/it/scala/stellar/sdk/TestNetworkJourney.scala
@@ -14,7 +14,7 @@ class TestNetworkJourney(implicit ee: ExecutionEnv) extends Specification with B
 
   implicit val network = TestNetwork
 
-  private val testAccounts = new TestAccounts(4)
+  private val testAccounts = new TestAccounts(6)
 
   def beforeAll() = testAccounts.open()
   def afterAll() = testAccounts.close()
@@ -34,7 +34,7 @@ class TestNetworkJourney(implicit ee: ExecutionEnv) extends Specification with B
 
     "be able to fee bump a v0 transaction" >> {
       val List(senderKey, recipientKey) = testAccounts.take(2)
-      val sender = Await.result(network.account((senderKey)), 10.seconds)
+      val sender = Await.result(network.account(senderKey), 10.seconds)
       val payment = PaymentOperation(recipientKey.toAccountId, Amount.lumens(2))
       val signedTransaction = Transaction(sender, List(payment), NoMemo, TimeBounds.Unbounded, NativeAmount(100))
         .sign(senderKey)
@@ -43,6 +43,25 @@ class TestNetworkJourney(implicit ee: ExecutionEnv) extends Specification with B
       val bumpedTxn = parsedV0Txn.bumpFee(NativeAmount(500), recipientKey)
       val response = Await.result(bumpedTxn.submit(), 20.seconds)
       response.isSuccess must beTrue
+    }
+  }
+
+  "a signing request for a transaction" should {
+    "be generated, parsed and executed" >> {
+      val List(senderKey, recipientKey) = testAccounts.take(2)
+      val sender = Await.result(network.account(senderKey), 10.seconds)
+
+      // Recipient prepares a signing request for themselves to be paid.
+      val payment = PaymentOperation(recipientKey.toAccountId, Amount.lumens(2))
+      val transaction = Transaction(sender, List(payment), NoMemo, TimeBounds.Unbounded, NativeAmount(100))
+      val requestUrl = transaction.signingRequest.toUrl
+
+      // Sender parses the URL, signs the request and submits to the network
+      val request = TransactionSigningRequest(requestUrl)
+      val parsedTransaction = request.transaction
+      val response = network.submit(parsedTransaction.sign(senderKey))
+
+      response.map(_.isSuccess) must beTrue.await(0, 1.minute)
     }
   }
 }